# control/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplyRequest": {
        "additionalProperties": false,
        "properties": {
          "report_digest": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Report Digest",
            "type": "string"
          },
          "source_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Source Sha256",
            "type": "string"
          },
          "source_yaml": {
            "maxLength": 262144,
            "minLength": 1,
            "title": "Source Yaml",
            "type": "string"
          }
        },
        "required": [
          "source_yaml",
          "source_sha256",
          "report_digest"
        ],
        "title": "ApplyRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PreviewRequest": {
        "additionalProperties": false,
        "properties": {
          "source_yaml": {
            "maxLength": 262144,
            "minLength": 1,
            "title": "Source Yaml",
            "type": "string"
          }
        },
        "required": [
          "source_yaml"
        ],
        "title": "PreviewRequest",
        "type": "object"
      },
...
      "ResolveImportRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "overlays": {
            "additionalProperties": true,
            "title": "Overlays",
            "type": "object"
          }
        },
        "required": [
          "expected_revision",
          "overlays"
        ],
        "title": "ResolveImportRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Vonk Forge Control",
    "version": "1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/catalog/imports/workload_run": {
      "post": {
        "operationId": "applyWorkloadRunImport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Apply"
      }
    },
    "/api/v1/catalog/imports/workload_run/preview": {
      "post": {
        "operationId": "previewWorkloadRunImport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Preview"
      }
    },
...
    "/api/v1/catalog/recipes/{recipe_id}/resolve-import": {
      "post": {
        "operationId": "resolveWorkloadRunImport",
        "parameters": [
          {
            "in": "path",
            "name": "recipe_id",
            "required": true,
            "schema": {
              "title": "Recipe Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Resolve Imported"
      }
    },
...
  }
}